+++ TracyPanels/Wireframe.php
@@ -8,7 +8,7 @@ use \Tracy\Dumper;
  *
  * See https://tracy.nette.org/en/extensions for docs about Tracy panels.
  *
- * @version 0.1.0
+ * @version 0.2.0
  * @author Teppo Koivula <teppo@wireframe-framework.com>
  * @license Mozilla Public License v2.0 https://mozilla.org/MPL/2.0/
  */
@@ -33,7 +33,28 @@ class WireframePanel extends BasePanel {
      *
      * @var string
      */
-    private $icon = '<svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path fill="#354B60" d="M12.41 148.02l232.94 105.67c6.8 3.09 14.49 3.09 21.29 0l232.94-105.67c16.55-7.51 16.55-32.52 0-40.03L266.65 2.31a25.607 25.607 0 00-21.29 0L12.41 107.98c-16.55 7.51-16.55 32.53 0 40.04zm487.18 88.28l-58.09-26.33-161.64 73.27c-7.56 3.43-15.59 5.17-23.86 5.17s-16.29-1.74-23.86-5.17L70.51 209.97l-58.1 26.33c-16.55 7.5-16.55 32.5 0 40l232.94 105.59c6.8 3.08 14.49 3.08 21.29 0L499.59 276.3c16.55-7.5 16.55-32.5 0-40zm0 127.8l-57.87-26.23-161.86 73.37c-7.56 3.43-15.59 5.17-23.86 5.17s-16.29-1.74-23.86-5.17L70.29 337.87 12.41 364.1c-16.55 7.5-16.55 32.5 0 40l232.94 105.59c6.8 3.08 14.49 3.08 21.29 0L499.59 404.1c16.55-7.5 16.55-32.5 0-40z"/></svg>';
+    private $icon = '<svg aria-hidden="true" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><!-- Font Awesome Free 5.15.1 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free (Icons: CC BY 4.0, Fonts: SIL OFL 1.1, Code: MIT License) --><path fill="#354B60" d="M12.41 148.02l232.94 105.67c6.8 3.09 14.49 3.09 21.29 0l232.94-105.67c16.55-7.51 16.55-32.52 0-40.03L266.65 2.31a25.607 25.607 0 00-21.29 0L12.41 107.98c-16.55 7.51-16.55 32.53 0 40.04zm487.18 88.28l-58.09-26.33-161.64 73.27c-7.56 3.43-15.59 5.17-23.86 5.17s-16.29-1.74-23.86-5.17L70.51 209.97l-58.1 26.33c-16.55 7.5-16.55 32.5 0 40l232.94 105.59c6.8 3.08 14.49 3.08 21.29 0L499.59 276.3c16.55-7.5 16.55-32.5 0-40zm0 127.8l-57.87-26.23-161.86 73.37c-7.56 3.43-15.59 5.17-23.86 5.17s-16.29-1.74-23.86-5.17L70.29 337.87 12.41 364.1c-16.55 7.5-16.55 32.5 0 40l232.94 105.59c6.8 3.08 14.49 3.08 21.29 0L499.59 404.1c16.55-7.5 16.55-32.5 0-40z"/></svg>';
+
+    /**
+     * Help icon
+     *
+     * @var string
+     */
+    private $iconHelp = '<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><!-- Font Awesome Free 5.15.1 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license/free (Icons: CC BY 4.0, Fonts: SIL OFL 1.1, Code: MIT License) --><path fill="#354B60" d="M256 8C119.043 8 8 119.083 8 256c0 136.997 111.043 248 248 248s248-111.003 248-248C504 119.083 392.957 8 256 8zm0 448c-110.532 0-200-89.431-200-200 0-110.495 89.472-200 200-200 110.491 0 200 89.471 200 200 0 110.53-89.431 200-200 200zm107.244-255.2c0 67.052-72.421 68.084-72.421 92.863V300c0 6.627-5.373 12-12 12h-45.647c-6.627 0-12-5.373-12-12v-8.659c0-35.745 27.1-50.034 47.579-61.516 17.561-9.845 28.324-16.541 28.324-29.579 0-17.246-21.999-28.693-39.784-28.693-23.189 0-33.894 10.977-48.942 29.969-4.057 5.12-11.46 6.071-16.666 2.124l-27.824-21.098c-5.107-3.872-6.251-11.066-2.644-16.363C184.846 131.491 214.94 112 261.794 112c49.071 0 101.45 38.304 101.45 88.8zM298 368c0 23.159-18.841 42-42 42s-42-18.841-42-42 18.841-42 42-42 42 18.841 42 42z"/></svg>';
+
+    /**
+     * Panel body
+     *
+     * @var string
+     */
+    private $body = '';
+
+    /**
+     * Resources URL
+     *
+     * @var string
+     */
+    private $resURL = '';
 
     /**
      * Wireframe instance
@@ -72,31 +93,79 @@ class WireframePanel extends BasePanel {
         // Get Wireframe
         $this->wireframe = $this->wire('modules')->get('Wireframe');
 
-        // Heading
-        $out = "<h1>{$this->icon} {$this->label}</h1>";
+        // Set up resources URL
+        $this->resURL = $this->wire('config')->urls->get('Wireframe') . 'res/';
 
-        // Maximize toggle
-        $out .= '<span class="tracy-icons"><span class="resizeIcons"><a href="#" title="Maximize / Restore" onclick="tracyResizePanel(\'' . $this->className . '\')">+</a></span></span>';
+        // Return panel content
+        return parent::loadResources()
+            . "<h1>{$this->icon} {$this->label}</h1>"
+            . "<link rel='stylesheet' href='{$this->resURL}css/TracyPanel.css'>"
+            . "<span class='tracy-icons'><span class='resizeIcons'><a href='#' title='Maximize / Restore' onclick='tracyResizePanel(\"{$this->className}\")'>+</a></span></span>"
+            . "<div class='tracy-inner'>"
+            . $this->getPanelBody()
+            . "<script>
+            (function() {
+                const script = document.createElement('script');
+                script.src = '{$this->resURL}js/TracyPanel.js';
+                document.getElementById('tracy-debug-panel-WireframePanel').appendChild(script);
+            })();
+            </script>"
+            . $this->getPanelFooter()
+            . "</div>";
+    }
 
-        // Panel body
-        $out .= '<div class="tracy-inner">';
-        $out .= $this->getPanelConfig();
-        $out .= $this->getPanelController();
-        $out .= $this->getPanelControllerProps();
-        $out .= $this->getPanelView();
-        $out .= $this->getPanelViewData();
+    /**
+     * Get panel body
+     *
+     * @return string
+     */
+    private function getPanelBody(): string {
 
-        // Panel footer
-        $out .= \TracyDebugger::generatePanelFooter(
+        // Prepare panel content
+        $content = array_filter([
+            'data' => [
+                'label' => 'Data',
+                'value' => $this->getPanelConfig()
+                    . $this->getPanelController()
+                    . $this->getPanelControllerProps()
+                    . $this->getPanelView()
+                    . $this->getPanelViewData()
+            ],
+            'api' => [
+                'label' => 'API',
+                'value' => $this->getPanelAPI(),
+            ],
+        ], function($tab) {
+            return $tab['value'] !== '';
+        });
+
+        // Construct and return panel body
+        $tab_class = ' wireframe-tracy-tab-link--current';
+        foreach ($content as $key => $tab) {
+            $this->body .= "<a class='wireframe-tracy-tab-link{$tab_class}' href='#wireframe-tracy-tab-{$key}'>{$tab['label']}</a>";
+            $tab_class = '';
+        }
+        $tab_attrs = '';
+        foreach ($content as $key => $tab) {
+            $this->body .= "<div class='wireframe-tracy-tab' id='wireframe-tracy-tab-{$key}'{$tab_attrs}>{$tab['value']}</div>";
+            $tab_attrs = ' hidden';
+        }
+        return $this->body;
+    }
+
+    /**
+     * Get panel footer
+     *
+     * @param string $body
+     * @return string
+     */
+    private function getPanelFooter(string $body = ''): string {
+        return \TracyDebugger::generatePanelFooter(
             $this->name,
             \Tracy\Debugger::timer($this->name),
-            strlen($out . '</div>'),
+            strlen($this->body),
             null
         );
-        $out .= '</div>';
-
-        return parent::loadResources()
-            . $out;
     }
 
     /**
@@ -105,7 +174,7 @@ class WireframePanel extends BasePanel {
      * @return string
      */
     private function getPanelConfig(): string {
-        $out = '<p><em>Config settings passed to the Wireframe module. <a href="https://wireframe-framework.com/docs/configuration-settings/">Documentation for config settings.</a></em></p>';
+        $out = "<a class='wireframe-tracy-doc-link' target='_blank' href='https://wireframe-framework.com/docs/configuration-settings/' title='Config settings passed to the Wireframe module.'>{$this->iconHelp}</a>";
         $out .= $this->renderTable($this->wireframe->getConfig());
         return $this->renderPanelSection('config', 'Config', $out, false);
     }
@@ -116,7 +185,7 @@ class WireframePanel extends BasePanel {
      * @return string
      */
     private function getPanelController(): string {
-        $out = '<p><em>Current Controller object. <a href="https://wireframe-framework.com/docs/controllers/">Documentation for Controllers.</a></em></p>';
+        $out = "<a class='wireframe-tracy-doc-link' target='_blank' href='https://wireframe-framework.com/docs/controllers/' title='Current Controller object.'>{$this->iconHelp}</a>";
         $controller = $this->wireframe->getController() ?: null;
         $out .= $controller === null ? '<pre>null</pre>' : $this->renderTable([
             'class' => get_class($controller),
@@ -131,7 +200,7 @@ class WireframePanel extends BasePanel {
      * @return string
      */
     private function getPanelControllerProps(): string {
-        $out = '<p><em>Public methods exposed by the Controller class, also known as Controller props. <a href="https://wireframe-framework.com/docs/controllers/">Documentation for Controllers.</a></em></p>';
+        $out = "<a class='wireframe-tracy-doc-link' target='_blank' href='https://wireframe-framework.com/docs/controllers/' title='Public methods exposed by the Controller class, also known as Controller props.'>{$this->iconHelp}</a>";
         $controller = $this->wireframe->getController() ?: null;
         if ($controller === null) {
             $out .= '<pre>null</pre>';
@@ -161,7 +230,7 @@ class WireframePanel extends BasePanel {
      * @return string
      */
     private function getPanelView(): string {
-        $out = '<p><em>Current View object. <a href="https://wireframe-framework.com/docs/view/">Documentation for the View layer.</a></em></p>';
+        $out = "<a class='wireframe-tracy-doc-link' target='_blank' href='https://wireframe-framework.com/docs/view/' title='Current View object.'>{$this->iconHelp}</a>";
         $view = $this->wireframe->view;
         $out .= $this->renderTable([
             'page' => $view->getPage(),
@@ -183,9 +252,124 @@ class WireframePanel extends BasePanel {
      * @return string
      */
     private function getPanelViewData(): string {
-        $out = '<p><em>Data (variables) passed via the bootstrap file and/or the Controller class to the View. <a href="https://wireframe-framework.com/docs/view/">Documentation for the View layer.</a></em></p>';
+        $out = "<a class='wireframe-tracy-doc-link' target='_blank' href='https://wireframe-framework.com/docs/view/' title='Data (variables) passed via the bootstrap file and/or the Controller class to the View.'>{$this->iconHelp}</a>";
         $out .= $this->renderTable($this->wireframe->view->data());
         return $this->renderPanelSection('viewData', 'View Data', $out, false);
+    }
+
+    /**
+     * Get API panel section
+     *
+     * @return string
+     */
+    private function getPanelAPI(): string {
+
+        // API debugger is limited to superusers
+        if (!$this->wire('user')->isSuperuser()) {
+            return '';
+        }
+
+        // Wireframe API module needs to be installed
+        if (!$this->wire('modules')->isInstalled('WireframeAPI')) {
+            return '<p>Wireframe API module is not installed, API debug tool disabled.</p>';;
+        }
+        $api = $this->wire('modules')->get('WireframeAPI');
+
+        // Fetch and check API endpoints
+        $enabled_endpoints = $api->getEnabledEndpoints();
+        if (empty($enabled_endpoints)) {
+            $edit_url = $this->wire('modules')->getModuleEditUrl($api);
+            return "<p>In order to perform API queries you need to <a href='{$edit_url}'>enable at least one API endpoint</a>.</p>";
+        }
+
+        // API root
+        $api_root = $this->wire('modules')->getModuleEditUrl('WireframeAPI', false) . '&api_query=';
+        $out = $this->renderInput('API root', 'api_root', 'text', null, $api_root);
+
+        // API endpoint
+        $out .= $this->renderInput('Endpoint', 'endpoint', 'select', null, $enabled_endpoints);
+
+        // Page ID
+        $out .= $this->renderInput('Page ID', 'page_id', 'number', 'pages', 1);
+
+        // Component
+        $available_components = array_map(function($component) {
+            return basename($component, '.php');
+        }, glob($this->wireframe->getConfig()['paths']['components'] . '*.php'));
+        $out .= $this->renderInput('Component', 'component', 'select', 'components', $available_components);
+
+        // Partial
+        $available_partials = [];
+        if ($this->wire('view')) {
+            $partials = $this->wire('view')->getPartials();
+            $partials_path_size = strlen($partials->getPath());
+            $available_partials = array_map(function($partial) use ($partials_path_size) {
+                return substr($partial, $partials_path_size);
+            }, $partials->getFilenames(null, false));
+        }
+        $out .= $this->renderInput('Partial', 'partial', 'select', 'partials', $available_partials);
+
+        // Return format
+        $out .= $this->renderInput('Return format', 'return_format', 'select', 'components+pages', [
+            '',
+            'json',
+            'rendered',
+        ]);
+
+        // Arguments
+        $out .= $this->renderInput('Arguments', 'return_format', 'textarea', null, "{\n\t\"argument\": \"value\"\n}", 'js-wireframe-tracy-api-args');
+        $out .= '<p style="opacity: .9">You can provide arguments as JSON, in which case they will be passed to the API as GET param "api_args", or in URL format (param1=value1&amp;param2=value2) in which case they will be appended to the API GET request as is. Note that the default API root used by this debugger only supports JSON format arguments.</p>';
+
+        // Render form
+        $out = "<form class='wireframe-tracy-api-form' id='js-wireframe-tracy-api-form'>"
+            . $out
+            . "<div class='wireframe-tracy-api-code wireframe-tracy-api-code--break' id='js-wireframe-tracy-api-query'></div>"
+            . "<div class='wireframe-tracy-api-code' id='js-wireframe-tracy-api-response' tabindex=-1 hidden></div>"
+            . "<div class='wireframe-tracy-api-form-row'><input type='submit' value='Send request'></div>"
+            . "</form>";
+
+        return $out;
+    }
+
+    /**
+     * Render form input
+     *
+     * @param string $label
+     * @param string $name
+     * @param string $type
+     * @param string|null $endpoint
+     * @param array|string|null $value
+     * @param string $class
+     * @return string
+     */
+    private function renderInput(string $label, string $name, string $type, ?string $endpoint = null, $value = null, $class = 'js-wireframe-tracy-api-param'): string {
+        $out = "<label class='wireframe-tracy-api-form-row";
+        if ($endpoint !== null) {
+            $out .= " wireframe-tracy-api-form-row--endpoint";
+            $endpoint_names = explode('+', $endpoint);
+            foreach ($endpoint_names as $endpoint_name) {
+                $out .= " wireframe-tracy-api-form-row--endpoint-{$endpoint_name}";
+            }
+        }
+        $out .= "'"
+            . ($endpoint !== null ? " hidden" : "")
+            . ">"
+            . "<span>{$label}</span>";
+        if ($type == 'number') {
+            $out .= "<input type='number' min=1 step=1 name='{$name}' value='{$value}' class='{$class}'>";
+        } else if ($type == 'text') {
+            $out .= "<input type='text' name='{$name}' value='{$value}' class='{$class}'>";
+        } else if ($type == 'textarea') {
+            $out .= "<textarea name='{$name}' class='{$class}' rows=5>{$value}</textarea>";
+        } else if ($type == 'select') {
+            $out .= "<select name='{$name}' id='js-wireframe-tracy-api-{$name}' class='{$class}'>"
+                . implode(array_map(function($value) {
+                    return "<option value='{$value}'>{$value}</option>";
+                }, $value))
+                . "</select>";
+        }
+        $out .= "</label>";
+        return $out;
     }
 
     /**
@@ -201,10 +385,10 @@ class WireframePanel extends BasePanel {
         if ($content === null) return '';
         $id = $this->wire('sanitizer')->name($id);
         $label = $this->wire('sanitizer')->entities($label);
-        return "<a href='#' rel='{$id}' class='tracy-toggle" . ($collapsed ? " tracy-collapsed" : "") . "'>"
+        return "<a href='#' rel='wireframe-tracy-panel-section-{$id}' class='tracy-toggle" . ($collapsed ? " tracy-collapsed" : "") . "'>"
             . $label
             . "</a>"
-            . "<div id='{$id}'" . ($collapsed ? " class='tracy-collapsed'" : "") . ">{$content}</div>"
+            . "<div class='wireframe-tracy-panel-section' id='wireframe-tracy-panel-section-{$id}'" . ($collapsed ? " class='tracy-collapsed'" : "") . ">{$content}</div>"
             . "<br>";
     }