+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Wireframe API debugger for Tracy, currently only available for superusers.
+- New method `Partial::getFilename(?string $ext = null, bool $with_ext = true)`. This is otherwise the same as type juggling a Partial object to string, but provides more customizability via params when called directly.
+- New methods `Partials::getPath()` and `Partials::getFilenames(string $ext = null, bool $with_ext = true)`.
+
+### Changed
+- Wireframe API built-in endpoints now provide more verbose errors if debug mode is enabled or user is logged in as superuser.
+
+### Fixed
+- In some cases Wireframe API wasn't able to catch errors; this part should now work more consistently.
+- Enabled endpoints setting in Wireframe API module config now works as expected.
+
 ## [0.18.2] - 2020-11-28
 
 ### Fixed